[ISSUE-310]  Added ignore for uuids in security group rules
Problem?

The uuid returned from an existing rule was causing all security
rules to show as if they had changes even when no changes to the rules
would be made.

Solution!

Added an `ignore_uuid` flag to the `apply_state` helper so that we would
be able to selectively ignore the uuid for certain objects, like
security group rules, where we don't yet support uuid as a parameter.