ci: run sanity checks with next release version set (#624)
rc job is executed on push to branches which runs semantic-release dry-run to pass the next version information to the sanity job, and sanity checks are run with version files updated temporarily without a commit. Sanity checks are still run on the same code version for PRs.

It also has a fix for ruff E402 check for eda logs.py file.